Wprowadzenie getopt do main.c, poprawienie Makefile oraz dodanie opcji clean - do usuwania zbędnych plików wywołania

+++ Makefile
@@ -1,11 +1,9 @@
 all:
 	cc board.c highscore.c levels.c sweeper.c main.c -o game
-	./game
-	rm game
 debug:
-	cc -Wall -pedantic board.c highscore.c levels.c sweeper.c main.c -o debug_mode 
-	./debug_mode
-	rm debug_mode
+	cc -Wall -pedantic board.c highscore.c levels.c sweeper.c main.c -o game
+clean:
+	@[ -f ./game ] && rm game || true
